refactor(app): move RobotSettingsDashboard
