+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -30,7 +30,7 @@ import { NameRobot } from '../pages/NameRobot'
 import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
 import { ProtocolSetup } from '../pages/ODD/ProtocolSetup'
 import { RobotDashboard } from '../pages/ODD/RobotDashboard'
-import { RobotSettingsDashboard } from '../pages/RobotSettingsDashboard'
+import { RobotSettingsDashboard } from '../pages/ODD/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../pages/ODD/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
 import { QuickTransferFlow } from '../organisms/QuickTransferFlow'
