+++ app/src/App/__tests__/OnDeviceDisplayApp.test.tsx
@@ -12,7 +12,7 @@ import { ConnectViaWifi } from '../../pages/ConnectViaWifi'
 import { NetworkSetupMenu } from '../../pages/NetworkSetupMenu'
 import { InstrumentsDashboard } from '../../pages/ODD/InstrumentsDashboard'
 import { RobotDashboard } from '../../pages/ODD/RobotDashboard'
-import { RobotSettingsDashboard } from '../../pages/RobotSettingsDashboard'
+import { RobotSettingsDashboard } from '../../pages/ODD/RobotSettingsDashboard'
 import { ProtocolDashboard } from '../../pages/ODD/ProtocolDashboard'
 import { ProtocolSetup } from '../../pages/ODD/ProtocolSetup'
 import { ProtocolDetails } from '../../pages/ODD/ProtocolDetails'
@@ -53,7 +53,7 @@ vi.mock('../../pages/ConnectViaEthernet')
 vi.mock('../../pages/ConnectViaUSB')
 vi.mock('../../pages/ConnectViaWifi')
 vi.mock('../../pages/ODD/RobotDashboard')
-vi.mock('../../pages/RobotSettingsDashboard')
+vi.mock('../../pages/ODD/RobotSettingsDashboard')
 vi.mock('../../pages/ODD/ProtocolDashboard')
 vi.mock('../../pages/ODD/ProtocolSetup')
 vi.mock('../../pages/ODD/ProtocolDetails')
