+++ app/src/organisms/RobotSettingsDashboard/DeviceReset.tsx
@@ -27,7 +27,7 @@ import { useDispatchApiRequest } from '../../redux/robot-api'
 
 import type { Dispatch, State } from '../../redux/types'
 import type { ResetConfigRequest } from '../../redux/robot-admin/types'
-import type { SetSettingOption } from '../../pages/RobotSettingsDashboard'
+import type { SetSettingOption } from '../../pages/ODD/RobotSettingsDashboard'
 import type { OddModalHeaderBaseProps } from '../../molecules/OddModal/types'
 
 interface LabelProps {
