+++ app/src/organisms/RobotSettingsDashboard/RobotSystemVersion.tsx
@@ -18,7 +18,7 @@ import { ChildNavigation } from '../../organisms/ChildNavigation'
 import { RobotSystemVersionModal } from './RobotSystemVersionModal'
 
 import type { RobotUpdateInfo } from '../../redux/robot-update/types'
-import type { SetSettingOption } from '../../pages/RobotSettingsDashboard'
+import type { SetSettingOption } from '../../pages/ODD/RobotSettingsDashboard'
 
 const GITHUB_URL = 'https://github.com/Opentrons/opentrons/releases'
